Rename homeworld panel modules to match sector and player modes.
File names said sector-only while the accordion and panel model also own the no-sector player-tile UX.
